Upgrade4920to41000: Migrate systemvmtemplate for 4.10
- Adds systemvm migration for upgrading to 4.10
- Remove systemvm migration block for upgrading from 4.5 or below, this
  ensures that pre 4.6 users are not required to seed 4.6 systemvmtemplate
  to upgrade to 4.10, but can directly seed a 4.10 systemvmtemplate
